requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4

# Environment Management
python-dotenv==1.0.0

# AI/ML
openai==2.15.0

# Production WSGI Server
gunicorn==21.2.0

# Security & CORS
flask-cors==4.0.0

# HTTP Requests (for health checks)
requests==2.31.0

# Legal RAG System
chromadb==0.4.22
pypdf==3.17.4
sentence-transformers==2.2.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
